Move GeneralEmail to its own service
