+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/service/GeneralEmailService.java
@@ -1,0 +1,10 @@
+package uk.gov.hmcts.reform.divorce.orchestration.service;
+
+import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CaseDetails;
+
+import java.util.Map;
+
+public interface GeneralEmailService {
+
+    Map<String, Object> createGeneralEmail(CaseDetails caseDetails) throws CaseOrchestrationServiceException;
+}
