+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/service/bulk/print/dataextractor/CaseDataExtractor.java
@@ -22,7 +22,7 @@ public class CaseDataExtractor {
         public static final String RESPONDENT_SOLICITOR_EMAIL = OrchestrationConstants.RESPONDENT_SOLICITOR_EMAIL_ADDRESS;
         public static final String CO_RESPONDENT_EMAIL = OrchestrationConstants.CO_RESP_EMAIL_ADDRESS;
         public static final String CO_RESPONDENT_SOLICITOR_EMAIL = CcdFields.CO_RESPONDENT_SOLICITOR_EMAIL_ADDRESS;
-        public static final String OTHER_PARTY_EMAIL = OrchestrationConstants.GENERAL_EMAIL_OTHER_RECIPIENT_EMAIL;
+        public static final String OTHER_PARTY_EMAIL = CcdFields.GENERAL_EMAIL_OTHER_RECIPIENT_EMAIL;
     }
 
     public static String getCaseReference(Map<String, Object> caseData) {
